feat: show files in a separate section on sidebar refactor: move file-related actions to a custom hook
