+++ src/lib/types.ts
@@ -1,0 +1,5 @@
+export enum ToastType {
+  Success = 'success',
+  Info = 'info',
+  Error = 'error',
+}
