+++ src/screens/SideMenu/Files.styled.ts
@@ -1,0 +1,6 @@
+import { SnIcon } from '@Components/SnIcon';
+import styled from 'styled-components/native';
+
+export const SNIconStyled = styled(SnIcon)`
+  margin-top: 10px;
+`;
