+++ src/screens/UploadedFilesList/UploadedFilesList.tsx
@@ -2,18 +2,15 @@ import { SearchBar } from '@Components/SearchBar';
 import { SnIcon } from '@Components/SnIcon';
 import { useNavigation } from '@react-navigation/native';
 import { ApplicationContext } from '@Root/ApplicationContext';
+import { useFiles } from '@Root/hooks/useFiles';
 import { ModalStackNavigationProp } from '@Root/ModalStack';
 import { SCREEN_UPLOADED_FILES_LIST } from '@Screens/screens';
 import { UploadedFileItem } from '@Screens/UploadedFilesList/UploadedFileItem';
 import {
-  UploadedFileItemAction,
-  UploadedFileItemActionType,
-} from '@Screens/UploadedFilesList/UploadedFileItemAction';
-import {
   UploadFilesListContainer,
   useUploadedFilesListStyles,
 } from '@Screens/UploadedFilesList/UploadedFilesList.styled';
-import { ChallengeReason, ContentType, SNFile } from '@standardnotes/snjs';
+import { SNFile } from '@standardnotes/snjs';
 import React, {
   FC,
   useCallback,
@@ -23,12 +20,9 @@ import React, {
   useRef,
   useState,
 } from 'react';
-import { FlatList, ListRenderItem, Platform, Text, View } from 'react-native';
-import RNFS, { exists } from 'react-native-fs';
+import { FlatList, ListRenderItem, Text, View } from 'react-native';
 import IosSearchBar from 'react-native-search-bar';
 import AndroidSearchBar from 'react-native-search-box';
-import RNShare from 'react-native-share';
-import Toast from 'react-native-toast-message';
 import { ThemeContext } from 'styled-components/native';
 
 enum Tabs {
@@ -45,10 +39,7 @@ export const UploadedFilesList: FC<Props> = props => {
   const navigation = useNavigation();
 
   const [currentTab, setCurrentTab] = useState(Tabs.AttachedFiles);
-  const [attachedFiles, setAttachedFiles] = useState<SNFile[]>([]);
-  const [allFiles, setAllFiles] = useState<SNFile[]>([]);
   const [searchString, setSearchString] = useState('');
-  const [isDownloading, setIsDownloading] = useState(false);
   const [filesListScrolled, setFilesListScrolled] = useState(false);
 
   const iosSearchBarInputRef = useRef<IosSearchBar>(null);
@@ -57,44 +48,7 @@ export const UploadedFilesList: FC<Props> = props => {
 
   const note = props.route.params.note;
 
-  const reloadAttachedFiles = useCallback(() => {
-    if (!application) {
-      return [];
-    }
-    setAttachedFiles(
-      application.items
-        .getFilesForNote(note)
-        .sort((a, b) => (a.created_at < b.created_at ? 1 : -1))
-    );
-  }, [application, note]);
-
-  const reloadAllFiles = useCallback(() => {
-    if (!application) {
-      return [];
-    }
-    setAllFiles(
-      application.items
-        .getItems(ContentType.File)
-        .sort((a, b) => (a.created_at < b.created_at ? 1 : -1)) as SNFile[]
-    );
-  }, [application]);
-
-  useEffect(() => {
-    if (!application) {
-      return;
-    }
-    const unregisterFileStream = application.streamItems(
-      ContentType.File,
-      () => {
-        reloadAttachedFiles();
-        reloadAllFiles();
-      }
-    );
-
-    return () => {
-      unregisterFileStream();
-    };
-  }, [application, reloadAllFiles, reloadAttachedFiles]);
+  const { attachedFiles, allFiles } = useFiles({ note });
 
   const filesList =
     currentTab === Tabs.AttachedFiles ? attachedFiles : allFiles;
@@ -152,320 +106,6 @@ export const UploadedFilesList: FC<Props> = props => {
     noAttachmentsIconContainer,
   } = styles;
 
-  /*const [selectedFile, setSelectedFile] = useState<DocumentPickerResponse[]>(
-    []
-  );
-  const handleDocumentSelection = async () => {
-    try {
-      const response = await DocumentPicker.pick({
-        presentationStyle: 'fullScreen',
-      });
-      console.log('file picked');
-      return response;
-    } catch (err) {
-      console.warn(err);
-    }
-  };*/
-
-  /*const handleError = (err: unknown) => {
-    if (DocumentPicker.isCancel(err)) {
-      console.warn('cancelled');
-      // User cancelled the picker, exit any dialogs or menus and move on
-    } else if (isInProgress(err)) {
-      console.warn(
-        'multiple pickers were opened, only the last will be considered'
-      );
-    } else {
-      throw err;
-    }
-  };*/
-
-  /*const uploadFile = async () => {
-    if (!application) {
-      return;
-    }
-    try {
-      const fileResult = await pickSingle(); // TODO: allow picking not only single (pickSingle), but multiple
-      if (!fileResult.uri || !fileResult.size) {
-        return;
-      }
-      const uri =
-        Platform.OS === 'ios' ? decodeURI(fileResult.uri) : fileResult.uri;
-      console.log(
-        'picked file',
-        fileResult.name,
-        fileResult.size,
-        fileResult.type,
-        uri
-      );
-
-      const size = fileResult.size;
-      const operation = await application.files.beginNewFileUpload();
-      const onChunk = async (
-        chunk: Uint8Array,
-        index: number,
-        isLast: boolean
-      ) => {
-        await application.files.pushBytesForUpload(
-          operation,
-          chunk,
-          index,
-          isLast
-        );
-      };
-
-      const data = await read(uri, size, 0, 'base64');
-      const bytes = Base64.toUint8Array(data);
-      await onChunk(bytes, 1, true);
-
-      /!*const fileObj = await application.files.finishUpload(
-        operation,
-        fileResult.name,
-        fileResult.type!
-      );*!/
-      const fileObj = await application.files.finishUpload(operation, {
-        name: fileResult.name,
-        mimeType: fileResult.type as string,
-      });
-
-      application.alertService.alert(
-        // `Uploaded file ${fileObj.nameWithExt}`
-        `Uploaded file ${fileObj.name}`
-      );
-    } catch (e) {
-      handleError(e);
-    }
-  };*/
-
-  const deleteFile = async (file: SNFile) => {
-    if (!application) {
-      return;
-    }
-    /*const shouldDelete = await confirmDialog({
-      text: `Are you sure you want to permanently delete "${file.nameWithExt}"?`,
-      confirmButtonStyle: 'danger',
-    });*/
-    const shouldDelete = await application.alertService.confirm(
-      `Are you sure you want to permanently delete "${file.name}"?`,
-      'Delete file',
-      'Delete'
-    );
-    if (shouldDelete) {
-      await application.files.deleteFile(file);
-      /*const deletingToastId = addToast({
-        type: ToastType.Loading,
-        message: `Deleting file "${file.nameWithExt}"...`,
-      });
-      await application.deleteItem(file);
-      addToast({
-        type: ToastType.Success,
-        message: `Deleted file "${file.nameWithExt}"`,
-      });
-      dismissToast(deletingToastId);*/
-    }
-  };
-
-  const deleteFileAtPath = async (path: string) => {
-    try {
-      if (await exists(path)) {
-        await RNFS.unlink(path);
-      }
-    } catch (err) {}
-  };
-
-  const shareFile = (file: SNFile, path: string) => {
-    application
-      .getAppState()
-      .performActionWithoutStateChangeImpact(async () => {
-        try {
-          const shareDialogResponse = await RNShare.open({
-            url: `file://${path}`,
-            failOnCancel: false,
-          });
-          // On Android this response always returns {success: false}, there is an open issue for that:
-          //  https://github.com/react-native-share/react-native-share/issues/1059
-          console.log('shareDialogResponse is', shareDialogResponse);
-
-          // On iOS the user can store files locally from "Share" screen, so we don't show "Download" option there.
-          // For Android the user has a separate "Download" action for the file, therefore after the file is shared,
-          // it's not needed anymore and we remove it from the storage.
-          if (Platform.OS === 'android') {
-            await deleteFileAtPath(path);
-          }
-          if (shareDialogResponse.success) {
-            Toast.show({
-              type: 'success',
-              text1: 'Successfully exported file',
-            });
-          }
-        } catch (error) {
-          Toast.show({
-            type: 'error',
-            text1: 'There was an error while sharing the file',
-          });
-        }
-      });
-  };
-
-  const downloadFile = async (file: SNFile, showShareScreen = false) => {
-    if (isDownloading) {
-      return;
-    }
-    const filesService = application.getFilesService();
-    const isGrantedStoragePermissionOnAndroid =
-      await filesService.hasStoragePermissionOnAndroid();
-
-    if (!isGrantedStoragePermissionOnAndroid) {
-      return;
-    }
-    setIsDownloading(true);
-
-    try {
-      Toast.show({
-        type: 'info',
-        text1: 'Downloading file...',
-        autoHide: false,
-      });
-
-      const path = filesService.getDestinationPath(file.name, showShareScreen);
-
-      // Overwrite any existing file with that name
-      await deleteFileAtPath(path);
-
-      await filesService.downloadFileInChunks(file, path);
-
-      Toast.hide();
-
-      if (showShareScreen) {
-        await shareFile(file, path);
-        return;
-      }
-
-      Toast.show({
-        type: 'success',
-        text1: 'Success',
-        text2: 'Successfully downloaded file',
-      });
-    } catch (error) {
-      Toast.show({
-        type: 'error',
-        text1: 'Error',
-        text2: 'There was an error while downloading the file',
-      });
-    } finally {
-      setIsDownloading(false);
-    }
-  };
-
-  const attachFileToNote = async (file: SNFile) => {
-    await application.items.associateFileWithNote(file, note);
-    Toast.show({
-      type: 'success',
-      text1: 'Successfully attached file to note',
-    });
-  };
-
-  const detachFileFromNote = async (file: SNFile) => {
-    await application.items.disassociateFileWithNote(file, note);
-    Toast.show({
-      type: 'success',
-      text1: 'Successfully detached file from note',
-    });
-  };
-
-  const toggleFileProtection = async (file: SNFile) => {
-    try {
-      if (!application) {
-        return file.protected;
-      }
-      let result: SNFile | undefined;
-      if (file.protected) {
-        result = await application.mutator.unprotectFile(file);
-      } else {
-        result = await application.mutator.protectFile(file);
-      }
-      const isProtected = result ? result.protected : file.protected;
-      return isProtected;
-    } catch (error) {
-      console.log('An error occurred: ', error);
-      return file.protected;
-    }
-  };
-
-  const authorizeProtectedActionForFile = async (
-    file: SNFile,
-    challengeReason: ChallengeReason
-  ) => {
-    if (!application) {
-      return false;
-    }
-    const authorizedFiles =
-      await application.protections.authorizeProtectedActionForFiles(
-        [file],
-        challengeReason
-      );
-    return authorizedFiles.length > 0 && authorizedFiles.includes(file);
-  };
-
-  const renameFile = async (file: SNFile, fileName: string) => {
-    await application.items.renameFile(file, fileName);
-  };
-
-  const handleFileAction = async (action: UploadedFileItemAction) => {
-    if (!application) {
-      return false;
-    }
-
-    const file =
-      action.type !== UploadedFileItemActionType.RenameFile
-        ? action.payload
-        : action.payload.file;
-    let isAuthorizedForAction = true;
-
-    if (
-      file.protected &&
-      action.type !== UploadedFileItemActionType.ToggleFileProtection
-    ) {
-      isAuthorizedForAction = await authorizeProtectedActionForFile(
-        file,
-        ChallengeReason.AccessProtectedFile
-      );
-    }
-
-    if (!isAuthorizedForAction) {
-      return false;
-    }
-
-    switch (action.type) {
-      case UploadedFileItemActionType.AttachFileToNote:
-        await attachFileToNote(file);
-        break;
-      case UploadedFileItemActionType.DetachFileToNote:
-        await detachFileFromNote(file);
-        break;
-      case UploadedFileItemActionType.DeleteFile:
-        await deleteFile(file);
-        break;
-      case UploadedFileItemActionType.ShareFile:
-        await downloadFile(file, true);
-        break;
-      case UploadedFileItemActionType.DownloadFile:
-        await downloadFile(file);
-        break;
-      case UploadedFileItemActionType.ToggleFileProtection: {
-        const isProtected = await toggleFileProtection(file);
-        action.callback(isProtected);
-        break;
-      }
-      case UploadedFileItemActionType.RenameFile:
-        await renameFile(file, action.payload.name);
-        break;
-    }
-
-    application.sync.sync();
-    return true;
-  };
-
   const onScroll = () => {
     if (filesListScrolled) {
       return;
@@ -478,8 +118,8 @@ export const UploadedFilesList: FC<Props> = props => {
       <UploadedFileItem
         key={item.uuid}
         file={item}
+        note={note}
         isAttachedToNote={attachedFiles.includes(item)}
-        handleFileAction={handleFileAction}
       />
     );
   };
@@ -520,7 +160,7 @@ export const UploadedFilesList: FC<Props> = props => {
           <View style={noAttachmentsIconContainer}>
             <SnIcon
               type={'files-illustration'}
-              styles={noAttachmentsIcon}
+              style={noAttachmentsIcon}
               width={72}
               height={72}
             />
